claude-code/composio-cli-agent/requirements.txt
composio-core>=0.5.0
typer>=0.12.0
python-dotenv>=1.0.0
mcp>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
httpx>=0.27.0